Adds option to start dependent processes
To specify starting/stopping dependent processes, the “client”
application’s test suite can be started with `$ DEP_SERVICES=auto_start
bundle exec spec`. If the environment variable `DEP_SERVICES=auto_start
` is not set, then the dependent services have to started/stopped
manually.

+++ client/spec/spec_helper.rb
@@ -57,28 +57,36 @@ RSpec.configure do |config|
   end
 
   config.before(:suite) do
-    Bundler.with_clean_env do
-      Dir.chdir(HOME_SRC_DIR) do
-        puts "[HOME] Starting test HOME server..."
-        `./bin/rails server --daemon --environment=test --pid=#{ HOME_PID } --port=#{ RemoteFactoryGirl.config.home[:port] }`
+    if auto_start_dependent_services?
+      Bundler.with_clean_env do
+        Dir.chdir(HOME_SRC_DIR) do
+          puts "[HOME] Starting test HOME server..."
+          `./bin/rails server --daemon --environment=test --pid=#{ HOME_PID } --port=#{ RemoteFactoryGirl.config.home[:port] }`
 
-        puts "[HOME] Preparing test HOME database..."
-        `bin/rake db:setup RAILS_ENV=test`
+          puts "[HOME] Preparing test HOME database..."
+          `bin/rake db:setup RAILS_ENV=test`
+        end
       end
     end
   end
 
   config.after(:suite) do
-    puts "\n[HOME] Stopping test HOME server..."
-    `cat #{ HOME_PID } | xargs kill -QUIT`
+    if auto_start_dependent_services?
+      puts "\n[HOME] Stopping test HOME server..."
+      `cat #{ HOME_PID } | xargs kill -QUIT`
 
-    Bundler.with_clean_env do
-      Dir.chdir(HOME_SRC_DIR) do
-        puts "[HOME] Destroying test HOME database..."
-        `bin/rake db:reset RAILS_ENV=test`
+      Bundler.with_clean_env do
+        Dir.chdir(HOME_SRC_DIR) do
+          puts "[HOME] Destroying test HOME database..."
+          `bin/rake db:reset RAILS_ENV=test`
+        end
       end
     end
   end
+end
+
+def auto_start_dependent_services?
+  ENV['DEP_SERVICES'] == 'auto_start'
 end
 
 #### Helpers used to help prove RemoteFactoryGirl created test data #####
